chardev: Let IOCanReadHandler use size_t type
Suggested-by: Paolo Bonzini <pbonzini@redhat.com>

+++ chardev/char-mux.c
@@ -190,7 +190,7 @@ static void mux_chr_accept_input(Chardev *chr)
     }
 }
 
-static int mux_chr_can_read(void *opaque)
+static size_t mux_chr_can_read(void *opaque)
 {
     MuxChardev *d = MUX_CHARDEV(opaque);
     int m = d->focus;
